aider: Modified the `incrementValue` and `decrementValue` functions in `public\comps.js` to accept `compsData` as a parameter and updated the calls to these functions to pass `compsData` as an argument.
# Aider chat conversation:

USER: `comps.js:870 incrementValue: compsData is undefined or compsUnitMixPct is not set. undefined
window.incrementValue @ comps.js:870
onclick @ (index):1
comps.js:884 decrementValue: compsData is undefined or compsUnitMixPct is not set. undefined`

ASSISTANT: The error message you're seeing suggests that the `compsData` object is not defined at the time when `incrementValue` or `decrementValue` functions are called. This could be due to the `compsData` not being properly set or being out of scope for the functions.

To resolve this issue, we need to ensure that `compsData` is available and properly defined within the scope of the `incrementValue` and `decrementValue` functions. One way to do this is to pass `compsData` as an argument to these functions when they are called.

Let's modify the `incrementValue` and `decrementValue` functions in `public\comps.js` to accept `compsData` as a parameter.

Here are the *SEARCH/REPLACE* blocks:

public\comps.js
```javascript
<<<<<<< SEARCH
window.incrementValue = function(button) {
=======
window.incrementValue = function(button, compsData) {
>>>>>>> REPLACE
```

public\comps.js
```javascript
<<<<<<< SEARCH
window.decrementValue = function(button) {
=======
window.decrementValue = function(button, compsData) {
>>>>>>> REPLACE
```

Additionally, we need to update the calls to `incrementValue` and `decrementValue` to pass `compsData` as an argument.

public\comps.js
```javascript
<<<<<<< SEARCH
<button class="increment-btn" onclick="incrementValue(this)">+</button>
=======
<button class="increment-btn" onclick="incrementValue(this, compsData)">+</button>
>>>>>>> REPLACE
```

public\comps.js
```javascript
<<<<<<< SEARCH
<button class="decrement-btn" onclick="decrementValue(this)">-</button>
=======
<button class="decrement-btn" onclick="decrementValue(this, compsData)">-</button>
>>>>>>> REPLACE
```

These changes will ensure that `compsData` is passed to the `incrementValue` and `decrementValue` functions, which should resolve the error you're encountering. Make sure to update all occurrences of these function calls in the HTML or JavaScript where they are used.